refactor component names
* setup test labels
* switch selected workout date to search query
