+++ root/frontend/src/components/Day/Day.tsx
@@ -11,7 +11,7 @@ interface PropTypes {
 }
 
 const isSameDate = (firstDate: Date, secondDate: Date): boolean => format(firstDate, 'MM-dd-yyyy') === format(secondDate, 'MM-dd-yyyy');
-const generateDateURL = (date: Date): string => `../Workouts/${format(date, 'MMddyyyy')}`;
+const generateDateURL = (date: Date): string => `../Workouts?date=${format(date, 'MMddyyyy')}`;
 
 function Day({ isSelected, date }: PropTypes) {
   const styles = useStyles();
